% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{calc_highquality}
\alias{calc_highquality}
\title{Calculate quantiles and identify high-quality pixels for species raster}
\usage{
calc_highquality(x, y = NULL, z = NULL, wah = FALSE, sb = FALSE)
}
\arguments{
\item{x}{A caribou calving resource selection function (RSF) or shorebird
habitat suitability index (HSI) RasterLayer object.}

\item{y}{Western Arctic Herd (WAH) caribou habitat overlap weighting
RasterLayer object (optional).}

\item{z}{Numeric value indicating the species-specific threshold value
defining shorebird suitable habitat from habitat suitability values (from
Saalfeld et al. 2013). Optional, only used if \code{sb == TRUE}.}

\item{wah}{Logical indicator of whether the run is for the WAH. If so, adds
the additional weighting analysis. Defaults to \code{FALSE}.}

\item{sb}{Logical indicator of whether the run is for a shorebird. If so,
adds a calculation of suitable habitat before calculating quantiles.
Defaults to \code{FALSE}.}
}
\value{
data.frame object containing the 75\% quantile value (\code{quant75}),
number of high-quality pixels in the input raster (\code{highquality.orig}),
and (if \code{wah = TRUE}) the weighted high-quality value (\code{highquality.weighted}).
}
\description{
\code{calc_highquality} is a helper function that prepares input data for the
caribou and shorebird impact analyses (\code{\link{impact_caribou}} and
\code{\link{impact_shorebird}}) by calculating quantiles of an input raster
and determining the number of high-quality pixels (\emph{sensu} Johnson et
al. 2005). See Fullman et al. (in review) for details.
}
\references{
Fullman TJ, Sullender BK, Cameron MD, Joly K. in review. Simulation modeling
accounts for uncertainty while quantifying ecological effects of development
alternatives. Ecosphere.

Johnson CJ, Boyce MS, Ray CL, Cluff DH, Gau RJ, Gunn A, Mulders R. 2005.
Cumulative effects of human developments on arctic wildlife. Wildlife
Monographs 160:1-36.

Saalfeld ST, Lanctot RB, Brown SC, Saalfeld DT, Johnson JA, Andres BA, Bart JR.
2013. Predicting breeding shorebird distributions on the Arctic Coastal
Plain of Alaska. Ecosphere 4:16.
}
