% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downstream_milo}
\alias{downstream_milo}
\title{Downstream data for Milo}
\format{
A dataframe with 925 observations of 4 variables
\describe{
  \code{flow_cfs } Value of flow (discharge)
  \code{cumulative_prob } Cumulative flow probability
  \code{flow_prob } Probability of individual flow value
  \code{smolt_surv_prob } Smolt survival probability given flow
}
}
\source{
Nieland et al. (2013, 2015)
}
\usage{
downstream_milo
}
\description{
Alden research estimates of downstream survival from the 
Atlantic salmon Dam Impact analysis model v67. Estimates are based on 
in-river flow.
}
\keyword{datasets}
