% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{footprint_generation}
\alias{footprint_generation}
\title{Convert point/line infrastructure to spatial footprints}
\usage{
footprint_generation(
  cpf.sf,
  sat.sf,
  rd.sf,
  proj.info,
  area.cpf = 100 * 4046.86,
  area.sat = 15 * 4046.86,
  road.width = 18.85953
)
}
\arguments{
\item{cpf.sf}{\code{sf POINT} object indicating central processing
facility (CPF) locations. Created by \code{\link{generate_cpf}}.}

\item{sat.sf}{\code{sf POINT} object indicating satellite production
facility locations. Created by \code{\link{generate_sat_rd}}.}

\item{rd.sf}{\code{SF LINESTRING} object indicating gravel road centerlines. Created
by \code{\link{generate_sat_rd}}.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{area.cpf}{Numeric value indicating the average size (sq. m) of a CPF
footprint. Default value equates to 100 acres based on BLM (2019b).}

\item{area.sat}{Numeric value indicating the average size (sq. m) of a
satellite production facility footprint. Default value equates to 15 acres based
on BLM (2019a) p.B-6.}

\item{road.width}{Numeric value indicating the estimated road ground
footprint width (m). Default value based on BLM (2019a) p.B-6. See Fullman
et al. (in press) for details.}
}
\value{
List containing three \code{sf POLYGON} objects: \code{cpf_footprints},
\code{sat_footprints}, and \code{surf_disturb}, which contain the CPF footprints,
satellite pad footprints, and combined footprint of CPFs, satellite pads, and roads,
respectively.
}
\description{
\code{footprint_generation} is a helper function that converts point and line
infrastructure locations, created by \code{\link{generate_cpf}} and \code{\link{generate_sat_rd}},
to spatial footprints of surface disturbance. This is used for each
species-specific impact analysis (\code{\link{impact_caribou}}, \code{\link{impact_shorebird}},
and \code{\link{impact_brant}}).
}
\references{
BLM [Bureau of Land Management] 2019a. National Petroleum Reserve
in Alaska Draft Integrated Activity Plan and Environmental Impact Statement.
Bureau of Land Management, U.S. Department of the Interior, Anchorage, AK,
USA.

BLM 2019b. Willow Master Development Plan Draft Environmental Impact
Statement. Bureau of Land Management, U.S. Department of the Interior.
Anchorage, AK, USA.

Fullman TJ, Sullender BK, Cameron MD, Joly K. in press. Simulation modeling
accounts for uncertainty while quantifying ecological effects of development
alternatives. Ecosphere.
}
