% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{gen_lcp_rd}
\alias{gen_lcp_rd}
\title{Road generation using least cost paths}
\usage{
gen_lcp_rd(tr, loc1, loc2, cur.cpf, cur.sat, proj.info)
}
\arguments{
\item{tr}{Scenario-specific \code{TransitionLayer} object, used by the \code{gdistance}
package for creating least cost paths.}

\item{loc1}{Numeric vector of length 2 indicating the x and y coordinates of the first location to be joined with a road.}

\item{loc2}{Numeric vector of length 2 indicating the x and y coordinates of the second location to be joined with a road.}

\item{cur.cpf}{Integer indicating the identity of the central processing facility (CPF) affiliated with the road to be generated.}

\item{cur.sat}{Integer indicating the identity of the satellite production pad affiliated with the road to be generated.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}
}
\value{
List of length 2 containing a \code{data.frame} with newly generated road
coordinates (\code{df_tmp}) and \code{sf LINESTRING} object depicting road locations
(\code{sf_tmp}).
}
\description{
\code{gen_lcp_rd} is a helper function that generates a road using least cost
paths and outputs the results as both a \code{data.frame} and \code{sf LINESTRING}
object. Used by \code{\link{lcp_rds_infield}}, \code{\link{lcp_rds_outfield}} via
\code{\link{gen_linkage_rd}}, and \code{\link{generate_sat_rd}}.
}
