% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{gen_linkage_rd}
\alias{gen_linkage_rd}
\title{Generate linkage road between anchor fields}
\usage{
gen_linkage_rd(
  x,
  y,
  w,
  v,
  road.res,
  tmp.rd.sf,
  tr,
  min.dist.exist,
  roadless.check,
  alt.b.stranded.leases = NULL,
  alt.d.TLnorth = NULL,
  scenario,
  proj.info,
  z,
  j
)
}
\arguments{
\item{x}{\code{data.frame} of locations to be connected to existing infrastructure.
For most scenarios this is the full anchor field dataset, \code{tmp.rd.df},
created by \code{\link{lcp_rds_infield}}, but \code{if(roadless.check == 2)}
it might be a subset of this \code{data.frame}, designated as \code{rd.0s}. This
is the same as the object \code{x} input in \code{\link{lcp_rds_outfield}}.}

\item{y}{\code{data.frame} of all anchor field road locations. This defaults to be
the same as \code{x}, which is valid for most scenarios. However,
\code{if(roadless.check == 2)} it can be specified as the full anchor field
dataset, \code{tmp.rd.df}, created by \code{\link{lcp_rds_infield}}, allowing
this to be passed along to the final dataset, even for roadless portions north
of Teshekpuk Lake that are not connected to existing infrastructure. This
is the same as the object \code{y} input in \code{\link{lcp_rds_outfield}}.}

\item{w}{\code{data.frame} of coordinates of existing/previously generated
infrastructure to which \code{x} will be connected.}

\item{v}{Road \code{sf LINESTRING} object to which the newly generated road will
be added. Set to \code{rd.exist} if \code{j == 1} and \code{rd.sf} otherwise.}

\item{road.res}{\code{SpatRaster} indicating resistance for least cost path road
generation. Created by \code{\link{prep_scenario_inputs}}.}

\item{tmp.rd.sf}{\code{sf LINESTRING} object depicting roads generated by
\code{\link{lcp_rds_infield}}.}

\item{tr}{Scenario-specific \code{TransitionLayer} object, used by the \code{gdistance}
package for creating least cost paths.}

\item{min.dist.exist}{Proximity order-sorted \code{data.frame} of distance from CPFs
to existing infrastructure. Created within \code{\link{generate_sat_rd}}.}

\item{roadless.check}{Integer value providing a flag that indicates how
roadless development should be dealt with. Created by \code{\link{generate_sat_rd}}.}

\item{alt.b.stranded.leases}{\code{SpatRaster} object identifying stranded lease locations
under Alternative B. Created by \code{\link{prep_scenario_inputs}}. Optional,
as this is only needed when running Alternative B.}

\item{alt.d.TLnorth}{\code{SpatRaster} object identifying available areas for development
north of Teshekpuk Lake under Alternative D. Created by \code{\link{prep_scenario_inputs}}.
Optional, as this is only needed when running Alternative D.}

\item{scenario}{Vector of character strings identifying the scenarios being run.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{z}{Iterator value used by \code{\link{dia}} to specify the development
scenario being analyzed.}

\item{j}{Iterator used within \code{\link{generate_sat_rd}} to indicate the
current CPF being analyzed.}
}
\value{
List containing two objects, a \code{data.frame} of newly generated linkage
road coordinates added to previously generated road coordinates (\code{rd_df})
and a \code{sf LINESTRING} object depicting all these roads (\code{rd_sf}).
These are fed into \code{\link{lcp_rds_outfield}}.
}
\description{
\code{gen_linkage_rd} is a helper function used by \code{\link{lcp_rds_outfield}}
to identify the closest point between a new anchor field (a central processing
facility and associated satellite production pads and connecting roads) and
previously generated/existing infrastructure and then link the two with a road
using least cost paths.
}
