% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{generate_sat_rd}
\alias{generate_sat_rd}
\title{Generate satellite pads and roads for each CPF}
\usage{
generate_sat_rd(
  cpf.df,
  cpf.sf,
  pad.exist,
  exist.coords,
  pad.prob,
  maxd2sat = 56327,
  mind2sat = 6437.4,
  n.sat = c(4, 8),
  n.iter = 100,
  alt.b.stranded.leases = NULL,
  tr.cost.alt.c.row = NULL,
  alt.d.TLnorth = NULL,
  tr.cost.alt,
  road.res,
  rd.exist,
  tr.cost.alt.b.stranded = NULL,
  wd.loc,
  path.out,
  proj.info,
  scenario,
  z,
  i,
  debug.out = FALSE
)
}
\arguments{
\item{cpf.df}{\code{data.frame} of CPF point locations. Created by \code{\link{generate_cpf}}.}

\item{cpf.sf}{\code{sf POINT} object depicting CPF point locations. Created by \code{\link{generate_cpf}}.}

\item{pad.exist}{\code{sf POINT} object of existing CPF and satellite pad locations.
Created by \code{\link{prep_general_inputs}}.}

\item{exist.coords}{\code{data.frame} containing coordinates of existing infrastructure.
Created by \code{\link{prep_general_inputs}}.}

\item{pad.prob}{\code{SpatRaster} layer depicting the scenario-specific relative
likelihood of CPF or satellite pad generation, relative to estimated undiscovered
oil, scenario-specific development restrictions, and waterbodies. Created by
\code{\link{prep_scenario_inputs}}.}

\item{maxd2sat}{Numeric value indicating the maximum allowed distance (m) from
a CPF to an associated satellite production pad. Defaults to 56,327 m based
on BLM (2019c) p.B-9.}

\item{mind2sat}{Numeric value indicating the minimum allowed distance (m)
between two satellite production pads, or a CPF and associated satellite
production pad. Defaults to 6437.4 m, similar to what is seen in the Willow
Master Development Plan (BLM 2019b).}

\item{n.sat}{Desired number of satellite pads to be simulated per CPF. This
can either be a fixed integer value or a range, \code{c(min,max)}, from
which a random value will be drawn each iteration using a uniform
distribution. Defaults to a range of \code{c(4, 8)} satellites per CPF.}

\item{n.iter}{Integer indicating the desired number of iterations to be run
for each scenario. Defaults to 100.}

\item{alt.b.stranded.leases}{\code{SpatRaster} object identifying stranded lease locations
under Alternative B. Created by \code{\link{prep_scenario_inputs}}. Optional,
as this is only needed when running Alternative B.}

\item{tr.cost.alt.c.row}{\code{TransitionLayer} for right-of-way (ROW) areas under
Alternative C, accounting for connecting roads that are allowed within some
of the 50\% Teshekpuk Caribou Herd (TCH) calving kernel under this alternative.
Created by \code{\link{prep_scenario_inputs}}. Optional, as this is only
needed when running Alternative C.}

\item{alt.d.TLnorth}{\code{SpatRaster} object identifying available areas for development
north of Teshekpuk Lake under Alternative D. Created by \code{\link{prep_scenario_inputs}}.
Optional, as this is only needed when running Alternative D.}

\item{tr.cost.alt}{Scenario-specific \code{TransitionLayer} indicating road
development cost. Created by \code{\link{prep_scenario_inputs}}.}

\item{road.res}{\code{SpatRaster} indicating resistance for least cost path road
generation. Created by \code{\link{prep_scenario_inputs}}.}

\item{rd.exist}{\code{sf LINESTRING} object representing existing roads.
Created by \code{\link{prep_general_inputs}}.}

\item{tr.cost.alt.b.stranded}{\code{TransitionLayer} specific to the stranded leases
in Alternative B. Created by \code{\link{prep_scenario_inputs}}. Optional, allowing
the code to be run for scenarios other than Alternative B.}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.out}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where output data will be saved. Defaults to
an "Output_Data" folder within the base folder.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{scenario}{Vector of character strings identifying the scenarios being run.}

\item{z}{Iterator value used by \code{\link{dia}} to specify the development
scenario being analyzed.}

\item{i}{Iterator value used by \code{\link{dia}} to run development simulation
and impacts analyses in parallel, indicating the specific iteration being run.}

\item{debug.out}{Logical indicator of whether intermediate infrastructure .csv
files should be written out for the purpose of debugging code. Defaults to
\code{FALSE}.}
}
\value{
List containing four objects: a \code{data.frame} of satellite pad locations
(\code{sat_df}), \code{sf POINT} object depicting satellite pad locations (\code{sat_sf}),
\code{data.frame} of road coordinates (\code{rd_df}), and \code{sf LINESTRING}
representation of road locations (\code{rd_sf}).
}
\description{
\code{generate_sat_rd} simulates the location of satellite production pads
as a function of relative estimated undiscovered oil, development restrictions,
and the presence of waterbodies. These are then connected to their parent
central processing facility (CPF) with roads using least cost paths, and the
entire "anchor facility" is connected via roads to existing or previously
simulated infrastructure.
}
\references{
BLM [Bureau of Land Management] 2019b. Willow Master Development
Plan Draft Environmental Impact Statement. Bureau of Land Management, U.S.
Department of the Interior. Anchorage, AK, USA.

BLM 2019c. Arctic National Wildlife Refuge Coastal Plain Oil and Gas Leasing
Program Final Environmental Impact Statement. Bureau of Land Management,
U.S. Department of the Interior, Anchorage, AK, USA.
}
