% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dam_passage.R
\name{get_dam_passage}
\alias{get_dam_passage}
\title{Get dam-specific survival rates}
\usage{
get_dam_passage(stoch_val = 0.1695, new_or_old = "new")
}
\arguments{
\item{stoch_val}{Value to be used for a stochastic random draw from 
`runif(-stoch_val, stoch_val)` to enforce variability on quantiles used
for flow-correlated survival values in the underlying data sets following
Nieland et al. (2020).}

\item{new_or_old}{A character string indicating whether to use `"new"` 
(Nieland et al. 2013, 2015) or `"old"` (Nieland et al. 2020) flow-correlated
probabilities of p_stillwater as well as flow-correlated survival at 
`milford`, `orono`, and `stillwater` dams.}
}
\value{
A vector with dam passage rates
}
\description{
Get correlated dam-specific downstream survival rates for 
smolts at each dam from built-in `downstream_` datasets.
}
\references{
Nieland et al. (2013, 2020).
}
