% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{impact_caribou}
\alias{impact_caribou}
\title{Caribou impact analysis}
\usage{
impact_caribou(
  surf.disturb,
  tch.raster = NULL,
  wah.raster = NULL,
  tch.calving = NULL,
  wah.calving = NULL,
  wah.hq.weight = NULL,
  out.df,
  wd.loc,
  path.out,
  proj.info,
  tch.hq = NULL,
  wah.hq = NULL,
  zoi = 4000,
  caribou.out = FALSE,
  scenario,
  n.iter = 100,
  z = NULL,
  i = NULL
)
}
\arguments{
\item{surf.disturb}{\code{sf POLYGON} object representing the combined
footprint of central processing facilities (CPFs), satellite pads, and
roads. Created by \code{\link{footprint_generation}}.}

\item{tch.raster}{Character string giving the file name of the Teshekpuk
Caribou Herd (TCH) calving raster. Optional, allowing the analysis to
be run for only one herd.}

\item{wah.raster}{Vector of two character strings, specifying the file names
of the Western Arctic Herd (WAH) calving raster and weighted overlap
raster, respectively. Optional, allowing the analysis to only be run
for one herd.}

\item{tch.calving, wah.calving}{\code{SpatRaster} object(s) depicting the calving resource
selection function (RSF) for the TCH and WAH, respectively. Optional,
allowing the analysis to be run for only one herd.}

\item{wah.hq.weight}{Western Arctic Herd (WAH) caribou habitat overlap
\code{SpatRaster} object (optional, unless analyses are being run for the WAH, in
which case it is required).}

\item{out.df}{\code{data.frame} object containing the scenario information and
infrastructure summary. Created during the impact analysis by \code{\link{dia}}.}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.out}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where output data will be saved. Defaults to
an "Output_Data" folder within the base folder.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{tch.hq, wah.hq}{\code{data.frame} object(s) created by \code{\link{calc_highquality}}.}

\item{zoi}{Numeric value indicating the zone of influence (m) in which habitat
quality is discounted. Defaults to 4000 m, following Wilson et al. (2013)
and Cameron et al. (2005).}

\item{caribou.out}{Logical indicator of whether the resulting discounted
caribou RSF raster(s) should be written out. Defaults to \code{FALSE}.}

\item{scenario}{Character vector of strings identifying the scenarios to be run.}

\item{n.iter}{Integer indicating the desired number of iterations to be run
for each scenario. Defaults to 100.}

\item{z}{Iterator value used by \code{\link{dia}} to specify the development
scenario being analyzed.}

\item{i}{Iterator value used by \code{\link{dia}} to run development simulation
and impacts analyses in parallel, indicating the specific iteration being run.}
}
\value{
\code{out.df data.frame}, updated to contain the caribou impact results.
}
\description{
\code{impact_caribou} runs the caribou impact analysis, discounting caribou
calving habitat values in proximity to infrastructure and calculating the
amount of "high quality" habitat remaining (\emph{sensu} Johnson et al. 2005).
}
\references{
Cameron RD, Smith WT, White RG, Griffith B. 2005. Central Arctic
caribou and petroleum development: Distributional, nutritional, and
reproductive implications. Arctic 58:1-9.

Johnson CJ, Boyce MS, Ray CL, Cluff DH, Gau RJ, Gunn A, Mulders R. 2005.
Cumulative effects of human developments on arctic wildlife. Wildlife
Monographs 160:1-36.

Wilson RR, Liebezeit JR, Loya WM. 2013. Accounting for uncertainty in oil and
gas development impacts to wildlife in Alaska. Conservation Letters
6:350-358.
}
