% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{impact_shorebird}
\alias{impact_shorebird}
\title{Shorebird impact analysis}
\usage{
impact_shorebird(sb.stack, rd.sl, sb.df)
}
\arguments{
\item{sb.stack}{RasterStack of shorebird habitat suitability index (HSI) rasters,
based on data from Saalfeld et al. (2013) and updated in the data preparation
code of \code{\link{dia}}.}

\item{rd.sl}{SpatialLines object indicating gravel road centerlines. Created
by \code{\link{generate_sat_rd}}.}

\item{sb.df}{data.frame containing the upper quantile of suitable high-quality
habitat values and the number of high-quality suitable pixels in the input
data. Created in the data preparation code of \code{\link{dia}}.}
}
\value{
\code{out.df} data.frame, updated to contain the shorebird impact results.
}
\description{
\code{impact_shorebird} runs the shorebird impact analysis, discounting
shorebird suitable habitat values in proximity to infrastructure and
calculating the amount of "high quality" suitable habitat remaining
(\emph{sensu} Johnson et al. 2005).
}
\references{
Johnson CJ, Boyce MS, Ray CL, Cluff DH, Gau RJ, Gunn A, Mulders R. 2005.
Cumulative effects of human developments on arctic wildlife. Wildlife
Monographs 160:1-36.

Saalfeld ST, Lanctot RB, Brown SC, Saalfeld DT, Johnson JA, Andres BA, Bart JR.
2013. Predicting breeding shorebird distributions on the Arctic Coastal
Plain of Alaska. Ecosphere 4:16.
}
