% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{in_river_m}
\alias{in_river_m}
\title{In river mortality}
\format{
A dataframe with 34 observations of 3 variables
\describe{
  \item{\code{mort_per_km}}{Production unit of origin}
  \item{\code{cum_freq}}{Cumulative probability at mort_per_km}
  \item{\code{probs}}{Sampling probability}
}
}
\source{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.
}
\usage{
in_river_m
}
\description{
Cumulative distribution function for in-river mortality of
outmigrating smolts per km.
}
\keyword{datasets}
