% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inefficiency_matrix}
\alias{inefficiency_matrix}
\title{Inefficiency matrix}
\format{
A matrix with 15 observations of 15 variables.
}
\source{
Nieland et al. (2013, 2015, 2020)
}
\usage{
inefficiency_matrix
}
\description{
Numeric columns 5 - 20 from \code{\link{upstream_inefficiency}}
}
\keyword{datasets}
