% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_summary.R
\name{inf_summary}
\alias{inf_summary}
\title{Summarize simulated infrastructure}
\usage{
inf_summary(
  out.df,
  cpf.spdf,
  sat.spdf,
  rd.sl,
  surf.disturb,
  npra.file,
  zoi = 4000,
  wd.loc,
  path.in
)
}
\arguments{
\item{out.df}{data.frame object containing the current scenario and iteration
information. Created during the impact analysis by \code{\link{dia}}.}

\item{cpf.spdf}{SpatialPointsDataFrame object indicating central processing
facility (CPF) locations. Created by \code{\link{generate_cpf}}.}

\item{sat.spdf}{SpatialPointsDataFrame object indicating satellite production
facility locations. Created by \code{\link{generate_sat_rd}}.}

\item{rd.sl}{SpatialLines object indicating gravel road centerlines. Created
by \code{\link{generate_sat_rd}}.}

\item{surf.disturb}{SpatialPolygonsDataFrame object representing the combined
footprint of central processing facilities (CPFs), satellite pads, and
roads. Created by \code{\link{footprint_generation}}.}

\item{npra.file}{Character string giving the file name of the NPR-A boundary
shapefile, without extension.}

\item{zoi}{Numeric value indicating the zone of influence (m) in which habitat
quality is discounted. Defaults to 4000 m, as is used in BLM (2019a).}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.in}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where input data are stored.}
}
\value{
\code{out.df} data.frame, updated to contain the infrastructure
summary results.
}
\description{
Infrastructure summary analysis. This calculates the number of facilities/length
of roads, area of direct surface disturbance, and area within 4 km of
infrastructure for comarison with statistics provided by the Bureau of Land
Management (BLM) in the Integrated Activity Plan Draft Environmental Impact
Statement (IAP DEIS) for the National Petroleum Reserve - Alaska (NPR-A)
(BLM 2019a). These metrics are added to the \code{out.df} data.frame.
}
\references{
BLM 2019a. National Petroleum Reserve in Alaska Draft Integrated
Activity Plan and Environmental Impact Statement. Bureau of Land Management,
U.S. Department of the Interior, Anchorage, AK, USA.
}
