% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{infrastructure_exclusion_buffer}
\alias{infrastructure_exclusion_buffer}
\title{Create infrastructure exclusion zones}
\usage{
infrastructure_exclusion_buffer(x, y, threshold)
}
\arguments{
\item{x}{Spatial object identifying the location(s) around which infrastructure
is to be excluded (i.e., around which raster values will be set to zero).}

\item{y}{\code{SpatRaster} object containing values that will be set to zero.}

\item{threshold}{Numeric value providing the distance (m) within which raster
values will be set to zero.}
}
\value{
\code{SpatRaster}, \code{y}, updated to have values within \code{threshold} of
infrastructure locations, \code{x}, set to zero.
}
\description{
\code{infrastructure_exclusion_buffer} is a helper function for infrastructure
generation that sets raster values to zero within a certain threshold of a
given feature (like a central processing facility) to ensure desired spacing
of infrastructure is maintained. Used by \code{\link{prep_general_inputs}}
and \code{\link{generate_sat_rd}}.
}
