% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{infrastructure_overlap_to_zero}
\alias{infrastructure_overlap_to_zero}
\title{Convert raster values to zero where they overlap a SpatialPolygons object}
\usage{
infrastructure_overlap_to_zero(x, surf.disturb)
}
\arguments{
\item{x}{Caribou calving resource selection function (RSF) RasterLayer object.}

\item{surf.disturb}{SpatialPolygonsDataFrame object representing the combined
footprint of central processing facilities (CPFs), satellite pads, and
roads. Created by \code{\link{footprint_generation}}.}
}
\value{
RasterLayer object with values of zero everywhere overlapped by
\code{surf.disturb}.
}
\description{
\strong{The \code{infrastructure_overlap_to_zero} is deprecated as of DIA version 0.2.0.
Please use \code{\link{raster_to_zero}} instead.}   \code{infrastructure_overlap_to_zero}
is a helper function for the caribou impact analysis (\code{\link{impact_caribou}})
that converts habitat values to 0 where they overlap the footprint of development.
}
