% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{infrastructure_proximity_discounting}
\alias{infrastructure_proximity_discounting}
\title{Discount raster values based on proximity to development and calculate remaining "high quality" habitat}
\usage{
infrastructure_proximity_discounting(
  x,
  surf.disturb,
  quant,
  wah.hq.weight = NULL,
  zoi = 4000,
  wah = FALSE,
  caribou.out = FALSE,
  proj.info,
  wd.loc,
  path.out = path.out,
  scenario,
  n.iter = 100,
  z = NULL,
  i = NULL
)
}
\arguments{
\item{x}{\code{SpatRaster} object indicating the caribou calving RSF values,
updated to set values overlapping the development footprint to zero. Created
using \code{\link{raster_to_zero}}.}

\item{surf.disturb}{\code{sf POLYGON} object representing the combined
footprint of central processing facilities (CPFs), satellite pads, and
roads. Created by \code{\link{footprint_generation}}.}

\item{quant}{Numerical value identifying the threshold for high-quality pixels of the calving
RSF raster. Defaults to the upper quartile of values. Calculated by \code{\link{calc_highquality}}.}

\item{wah.hq.weight}{Western Arctic Herd (WAH) caribou habitat overlap
\code{SpatRaster} object (optional, unless analyses are being run for the WAH, in
which case it is required).}

\item{zoi}{Numeric value indicating the zone of influence (m) in which habitat
quality is discounted. Defaults to 4000 m, following Wilson et al. (2013)
and Cameron et al. (2005).}

\item{wah}{Logical indicator of whether the run is for the WAH. If so, runs
the additional weighting analysis. Defaults to \code{FALSE}.}

\item{caribou.out}{Logical indicator of whether the resulting discounted
caribou RSF raster(s) should be written out. Defaults to \code{FALSE}.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.out}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where output data will be saved. Defaults to
an "Output_Data" folder within the base folder.}

\item{scenario}{Character vector of strings identifying the scenarios to be run.}

\item{n.iter}{Integer indicating the desired number of iterations to be run
for each scenario. Defaults to 100.}

\item{z}{Iterator value used by \code{\link{dia}} to specify the development
scenario being analyzed.}

\item{i}{Iterator value used by \code{\link{dia}} to run development simulation
and impacts analyses in parallel, indicating the specific iteration being run.}
}
\value{
Integer value specifying the number of high-quality habitat pixels
(weighted value if \code{wah = TRUE}) remaining after discounting.
}
\description{
\code{infrastructure_proximity_discounting} is a helper function for the
caribou impact analysis (\code{\link{impact_caribou}}) that discounts
resource selection function (RSF) values that do not directly overlap the
surface footprint of infrastructure based on their proximity to development.
Then calculates the amount of remaining "high quality" habitat (\emph{sensu}
Johnson et al. 2005).
}
\references{
Cameron RD, Smith WT, White RG, Griffith B. 2005. Central Arctic
caribou and petroleum development: Distributional, nutritional, and
reproductive implications. Arctic 58:1-9.

Johnson CJ, Boyce MS, Ray CL, Cluff DH, Gau RJ, Gunn A, Mulders R. 2005.
Cumulative effects of human developments on arctic wildlife. Wildlife
Monographs 160:1-36.

Wilson RR, Liebezeit JR, Loya WM. 2013. Accounting for uncertainty in oil and
gas development impacts to wildlife in Alaska. Conservation Letters
6:350-358.
}
