% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{infrastructure_spacer}
\alias{infrastructure_spacer}
\title{Simulate locations proportional to suitability and outside unavailable areas}
\usage{
infrastructure_spacer(n.dev, im, threshold)
}
\arguments{
\item{n.dev}{Integer value indicating the desired number of simulated
developments (e.g., central processing facilities or satellite pads).}

\item{im}{Object of class \code{im} that is used by spatstat to randomly
generate locations proportional to suitability and constrained to avoid
restrictions like water or areas closed to development.}

\item{threshold}{Numeric value indicating the minimum spacing (m) between
simulated infrastructure locations.}
}
\value{
\code{data.frame} containing the xy coordinates of the simulated development.
}
\description{
\code{infrastructure_spacer} is a helper function for infrastructure simulation
that generates infrastructure locations proportional to a suitability
raster (e.g., relative estimated undiscovered oil) and constrained to not
occur in unavailable areas or within a minimum distance threshold from other
infrastructure. Used by \code{\link{generate_cpf}} and \code{\link{generate_sat_rd}}.
}
