% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{lcp_rds_outfield}
\alias{lcp_rds_outfield}
\title{Connect anchor field to existing infrastructure}
\usage{
lcp_rds_outfield(
  x,
  y = x,
  tmp.rd.sf,
  rd.df = NULL,
  rd.sf = NULL,
  tr,
  min.dist.exist,
  exist.coords,
  road.res,
  rd.exist,
  alt.b.stranded.leases = NULL,
  tr.cost.alt.c.row = NULL,
  alt.d.TLnorth = NULL,
  wd.loc,
  path.out,
  scenario,
  n.iter,
  roadless.check,
  proj.info,
  j,
  z = NULL,
  i = NULL,
  debug.out = FALSE
)
}
\arguments{
\item{x}{\code{data.frame} of locations to be connected to existing infrastructure.
For most scenarios this is the full anchor field dataset, \code{tmp.rd.df},
created by \code{\link{lcp_rds_infield}}, but \code{if(roadless.check == 2)}
it might be a subset of this \code{data.frame}, designated as \code{rd.0s}.}

\item{y}{\code{data.frame} of all anchor field road locations. This defaults to be
the same as \code{x}, which is valid for most scenarios. However,
\code{if(roadless.check == 2)} it can be specified as the full anchor field
dataset, \code{tmp.rd.df}, created by \code{\link{lcp_rds_infield}}, allowing
this to be passed along to the final dataset, even for roadless portions north
of Teshekpuk Lake that are not connected to existing infrastructure.}

\item{tmp.rd.sf}{\code{sf LINESTRING} object depicting anchor field roads. Created by
\code{\link{lcp_rds_infield}}.}

\item{rd.df}{\code{data.frame} of road coordinates. If this is the first iteration
through \code{\link{generate_sat_rd}} (i.e., \code{j == 1}) then this will
not exist and will be created by the code. Otherwise, the file previously
created/updated by \code{lcp_rds_outfield} will be used.}

\item{rd.sf}{\code{sf LINESTRING} object depicting road locations. If this is the first
iteration through \code{\link{generate_sat_rd}} (i.e., \code{j == 1}) then
this will not exist and will be created by the code. Otherwise, the file
previously created/updated by \code{lcp_rds_outfield} will be used.}

\item{tr}{Scenario-specific \code{TransitionLayer} object, used by the \code{gdistance}
package for creating least cost paths.}

\item{min.dist.exist}{Proximity order-sorted \code{data.frame} of distance from CPFs
to existing infrastructure. Created within \code{\link{generate_sat_rd}}.}

\item{exist.coords}{\code{data.frame} containing coordinates of existing infrastructure.
Created by \code{\link{prep_general_inputs}}.}

\item{road.res}{Scenario-specific road restriction \code{SpatRaster}. Created by
\code{\link{prep_scenario_inputs}}.}

\item{rd.exist}{\code{sf LINESTRING} object representing existing roads.
Created by \code{\link{prep_general_inputs}}.}

\item{alt.b.stranded.leases}{\code{SpatRaster} object identifying stranded lease locations
under Alternative B. Created by \code{\link{prep_scenario_inputs}}. Optional,
as this is only needed when running Alternative B.}

\item{tr.cost.alt.c.row}{\code{TransitionLayer} for right-of-way (ROW) areas under
Alternative C, accounting for connecting roads that are allowed within some
of the 50\% Teshekpuk Caribou Herd (TCH) calving kernel under this alternative.
Created by \code{\link{prep_scenario_inputs}}. Optional, as this is only
needed when running Alternative C.}

\item{alt.d.TLnorth}{\code{SpatRaster} object identifying available areas for development
north of Teshekpuk Lake under Alternative D. Created by \code{\link{prep_scenario_inputs}}.
Optional, as this is only needed when running Alternative D.}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.out}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where output data will be saved. Defaults to
an "Output_Data" folder within the base folder.}

\item{scenario}{Vector of character strings identifying the scenarios being run.}

\item{n.iter}{Integer indicating the desired number of iterations to be run
for each scenario. Defaults to 100.}

\item{roadless.check}{Integer value providing a flag that indicates how
roadless development should be dealt with. Created by \code{\link{generate_sat_rd}}.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{j}{Iterator used within \code{\link{generate_sat_rd}} to indicate the
current CPF being analyzed.}

\item{z}{Iterator value used by \code{\link{dia}} to specify the development
scenario being analyzed.}

\item{i}{Iterator value used by \code{\link{dia}} to run development simulation
and impacts analyses in parallel, indicating the specific iteration being run.}

\item{debug.out}{Logical indicator of whether intermediate infrastructure .csv
files should be written out for the purpose of debugging code. Defaults to
\code{FALSE}.}
}
\value{
List containing two objects, a \code{data.frame} of generated road coordinates
(\code{rd_df}) and an \code{sf LINESTRING} object depicting generated roads (\code{rd_sf}).
}
\description{
\code{lcp_rds_outfield} is a helper function that operates within \code{\link{generate_sat_rd}}
to connect an anchor field (a central processing facility, or CPF, and its
associated satellites, connected by infield roads created by \code{\link{lcp_rds_infield}})
to existing or previously simulated infrastructure.
}
