% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_spatial}
\alias{load_spatial}
\title{Load spatial data and ensure projection matching}
\usage{
load_spatial(x, proj.info, vec = FALSE, wd.loc, path.in)
}
\arguments{
\item{x}{Character string specifying the file name containing the spatial object.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{vec}{Logical indicator of whether the spatial data to be loaded is in
vector format (\code{TRUE}) or raster data (\code{FALSE}; default).}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.in}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where input data are stored. Defaults to an
Input_Data folder within the base folder.}
}
\value{
A Raster* or Spatial* object, projected to the desired project-wide
projection.
}
\description{
\code{load_spatial} is a helper function that loads spatial data (vector or
raster) and ensures the projection information matches exactly that selected
for the overall analysis.
}
