% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs_per_female.R
\name{make_eggs_per_female}
\alias{make_eggs_per_female}
\title{Simulate eggs per female}
\usage{
make_eggs_per_female(n = 1, a = 4000, b = 12000, mean = NULL, sd = NULL)
}
\arguments{
\item{n}{Number of samples}

\item{a}{Lower limit of truncated normal}

\item{b}{Upper limit of truncated normal}

\item{mean}{Mean of distribution. When `mean = NULL` (default), the mean
value is parameterized from the built-in \code{\link{life_stage_survival}}
dataset.}

\item{sd}{Standard deviation of distribution. When `mean = NULL` (default), 
the mean value is parameterized from the built-in
 \code{\link{life_stage_survival}} dataset.}
}
\value{
A vector of length `n`
}
\description{
Simulate number of eggs per adult female based on data
collected at the US Fish and Wildlife Service Craig Brook National Fish
Hatchery.
}
\examples{
# 1. Simulate a single value for eggs per female ----
make_eggs_per_female()

# 2. Simulate 10,000 values for eggs per female ----
make_eggs_per_female(n = 10000)

}
\references{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.
}
