% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inriver_s.R
\name{make_inriver_s}
\alias{make_inriver_s}
\title{Create natural in-river smolt survival by reach}
\usage{
make_inriver_s(m = NULL, initial = FALSE, hatchery_wild = "wild", alt = 0)
}
\arguments{
\item{m}{Natural in-river mortality rate per km. The default (`NULL`) 
uses a truncated normal distribution with mean = 0.00329 and standard 
deviation of 0.00459 defined on the interval (0, 0.29)}

\item{initial}{Logical indicating Whether in-river mortality rate is for initial PU or
a PU downstream from where smolts started. If `TRUE` then default hatchery 
segment lengths is used to calculate whole-reach survival for hatchery smolts
as (1 - m) ^ dia::production_units$Hatchery_segment_length) and 1/2 of the
longest straight-line reach length 
(dia::production_units$Longest_segment_length_km) is used to calculate 
whole-reach survival of wild smolts. If `FALSE`, then both hatchery and wild
smolt survival is calculated from the longets straight-line reach length.}

\item{hatchery_wild}{A character string indicating whether reach mortality is
to be calculated for `"hatchery"` or `"wild"` smolts.}

\item{alt}{A numeric indicating whether to use 
`"Longest_segment_length"` (`alt = 0`), 
`"Partial_segment_length_1"` (`alt = 2`), or 
`"Partial_segment_length_2"` (`alt = 3`) from the `production_units` table to
calculate whole-reach survival rates based on per-km mortality.}
}
\value{
A vector with 15 elements, I think...
}
\description{
Create natural in-river smolt mortality per kilometer from 
built-in datasets.
}
\references{
Nieland et al. (2020).
}
