% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_dams}
\alias{n_dams}
\title{Number of dams by natal production unit (PU) and migration route}
\format{
A dataframe with 15 observations of 3 variables
\describe{
  \code{PU } Production unit of origin \cr \cr
  \code{stillwater } Number of dams by natal PU encountered during 
  downstream migration if fish move through the Stillwater Branch \cr \cr
  \code{mainstem } Number of dams by natal PU encountered during 
  downstream migration if fish move through the mainstem Penobscot River \cr \cr  
}
}
\source{
Nieland et al. (2020)
}
\usage{
n_dams
}
\description{
Number of dams passed by out-migrating smolts starting in
different production units by downstream migration route (`stillwater` or
`mainstem`)
}
\keyword{datasets}
