% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{poly_rotate}
\alias{poly_rotate}
\title{Polygon rotation}
\usage{
poly_rotate(a)
}
\arguments{
\item{a}{Numeric value indicating the desired amount of rotation, in degrees.}
}
\value{
A \code{data.frame} with the coordinates of the rotated polygon coordinates.
}
\description{
\code{poly_rotate} is a helper function that rotates a polygon object around
its centroid. Based on code from \href{https://geocompr.robinlovelace.net/geometric-operations.html}{Geocomputation with R}.
}
