% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{prep_scenario_inputs}
\alias{prep_scenario_inputs}
\title{Prepare input data needed for an individual scenario}
\usage{
prep_scenario_inputs(
  scenario,
  oil.av,
  cost.map,
  cost.water,
  willow.cpf.ras,
  pad.res.file,
  road.res.file,
  alt.b.rd.stranded.res.file = NULL,
  alt.b.stranded.lease.file = NULL,
  alt.c.row.file = NULL,
  alt.d.north.file = NULL,
  wd.loc,
  path.in,
  proj.info,
  z
)
}
\arguments{
\item{scenario}{Character vector of strings identifying the scenarios to be run.}

\item{oil.av}{\code{SpatRaster} indicating the relative estimated undiscovered oil,
created by \code{\link{prep_general_inputs}}.}

\item{cost.map}{\code{SpatRaster} indicating land and water areas, used in gravel
pad simulation. Created by \code{\link{prep_general_inputs}}.}

\item{cost.water}{\code{SpatRaster} indicating road-specific cost map where water
areas are assigned a value of 0.05 to indicate the possibility, with lower
likelihood, of water crossings (Wilson et al. 2013). Created by \code{\link{prep_general_inputs}}.}

\item{willow.cpf.ras}{\code{SpatRaster} that restricts central processing facility
(CPF) generation within \code{d2cpf} of the Willow CPF. Created by
\code{\link{prep_general_inputs}}.}

\item{pad.res.file}{Character string specifying the file name for the raster
containing scenario-specific restrictions for gravel pad (i.e., CPF and
satellite pad) placement. Restriction values may consist of 0 (no development),
0.1 (reduced likelihood of development), or 1 (unrestricted development).}

\item{road.res.file}{Character string indicating the file name for the raster
containing scenario-specific restrictions for road placement. Restriction values
may consist of 0 (no development), 0.1 (reduced likelihood of development), or
1 (unrestricted development).}

\item{alt.b.rd.stranded.res.file}{Character string specifying the file name for
the road development restriction raster for stranded leases under Alternative
B. Optional, allowing the code to be run for scenarios other than Alternative B.}

\item{alt.b.stranded.lease.file}{Character string specifying the file name for the
raster identifying stranded lease locations under Alternative B. Optional,
allowing the code to be run for scenarios other than Alternative B.}

\item{alt.c.row.file}{Character string specifying the file name for the raster
identifying right-of-way (ROW) areas under Alternative C where infield roads
are prohibited but connecting infrastructure is allowed. Optional, allowing
the code to be run for scenarios other than Alternative C.}

\item{alt.d.north.file}{Character string specifying the file name for the
raster identifying discrete areas north of Teshekpuk Lake where roadless
development is possible under Alternative D. Optional, allowing the code to
be run for scenarios other than Alternative D.}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.in}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where input data are stored. Defaults to an
Input_Data folder within the base folder.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}

\item{z}{Iterator value used by \code{\link{dia}} to specify the development
scenario being analyzed.}
}
\value{
List containing scenario-specific data for: \code{SpatRaster} indicating
the relative likelihood of pad development as a function of relative
estimated undiscovered oil, development restrictions, and the presence of
waterbodies (\code{oil_av}), \code{im} version of \code{oil_av} (\code{oil_im}),
\code{SpatRaster} indicating resistance for least cost path road generation
(\code{road_res}), \code{TransitionLayer} indicating road development cost
(\code{tr_cost_alt}), \code{TransitionLayer} specific to the stranded leases in
Alternative B (\code{tr_cost_B_stranded}), \code{SpatRaster} indicating location
of stranded leases in Alternative B (\code{alt_B_stranded_leases}),
\code{TransitionLayer} specific to the stranded leases in the ROW area south of
Teshekpuk Lake for Alternative C (\code{tr_cost_C_row}), \code{SpatRaster}
identifying discrete areas north of Teshekpuk Lake where roadless
development is possible under Alternative D (\code{alt_D_north}).
}
\description{
Some input data needed for infrastructure simulation is common to all scenarios
while other input data is scenario-specific. The former is prepared by
\code{\link{prep_general_inputs}}, while \code{prep_scenario_inputs} prepares
the scenario-specific data. It thus must be run separately for each scenario.
}
\references{
Wilson RR, Liebezeit JR, Loya WM. 2013. Accounting for uncertainty in oil and
gas development impacts to wildlife in Alaska. Conservation Letters
6:350-358.
}
