% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projection_alignment}
\alias{projection_alignment}
\title{Ensure matching projections of spatial data}
\usage{
projection_alignment(x, proj.info)
}
\arguments{
\item{x}{Raster* or Spatial* object.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}
}
\value{
Spatial object of the same class as \code{x}, projected to the
desired project-wide projection.
}
\description{
\code{projection_alignment} is a helper function that ensures the projection
information of a spatial object matches exactly that selected for the overall
analysis. It checks whether the projection of an object matches the desired
projection information, indicated by \code{proj.info} and, if not, projects
the object.
}
\details{
This function is intended to address discrepancies in spatial object
projection, not to deal with issues of misaligned extent or cell size for
Raster* objects. \code{projection_alignment} assumes that the user has already
done initial preparation to ensure any raster files have the same extent and
resolution. For tools to address such data preparation, see the \href{https://cran.r-project.org/web/packages/raster/index.html}{raster}
package.
}
