% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{pt_to_pad}
\alias{pt_to_pad}
\title{Gravel pad polygon creation}
\usage{
pt_to_pad(x, area, proj.info)
}
\arguments{
\item{x}{\code{sf POINT} object providing pad centroid location(s).}

\item{area}{Numeric value indicating the desired area of the final polygon object,
in square meters.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}
}
\value{
An \code{sf POLYGON} object representing the footprint of CPF or satellite
gravel pads.
}
\description{
\code{pt_to_pad} is a helper function that converts a point location to a polygon
by creating a square buffer of a desired area around the point (which is taken to
be the centroid). Used to represent gravel pad area of central processing facility
(CPF) and satellite production pads. Used by \link{footprint_generation}.
}
