% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raster_to_zero}
\alias{raster_to_zero}
\title{Set raster values to zero}
\usage{
raster_to_zero(x, y)
}
\arguments{
\item{x}{\code{SpatRaster} object.}

\item{y}{Polgon layer of class \code{SpatVector} representing the area within which
the raster values will be set to zero.}
}
\value{
\code{SpatRaster} object with the same information as \code{x}, but with
areas overlapping \code{y} set to zero.
}
\description{
\code{raster_to_zero} is a helper function that sets raster values to zero where
they are overlapped by a polygon. Pixels with an \code{NA} value in the input raster
will remain \code{NA} in the output.
}
\note{
\code{raster_to_zero} replaces the \code{infrastructure_overlap_to_zero}
function from \code{dia} version 0.1.0. It also simplifies the code for the
\code{infrastructure_exclusion_buffer} function from that version.
}
