% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dia_shiny.R
\name{run_dia_shiny}
\alias{run_dia_shiny}
\title{Run the DIA model interactively with shiny}
\usage{
run_dia_shiny()
}
\description{
Run the DIA model interactively in a browser using the shiny
package for R.
}
\seealso{
run_dia
}
