% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_downstream_passage.R
\name{run_downstream_passage}
\alias{run_downstream_passage}
\title{Run downstream passage routine}
\usage{
run_downstream_passage(
  stocked_smolts,
  wild_smolts,
  downstream_passage,
  p_stillwater,
  mattaceunk_impoundment_mortality,
  n_dams,
  indirect_latent_mortality,
  p_female,
  in_river_s
)
}
\arguments{
\item{stocked_smolts}{Number of hatchery smolts stocked by PU.}

\item{wild_smolts}{Number of wild-reared smolts.}

\item{downstream_passage}{Numeric vector with length = 15 corresponding to
dam passage efficiencies at each of 15 dams in the watershed that are used 
to delineate PUs. Default usage is to inherit these values from the 
\code{link{run_dia}} function.}

\item{p_stillwater}{Probability that fish use the Stillwater Branch for 
downstream migration. Default usage is to inherit this value from the 
\code{link{run_dia}} function.}

\item{mattaceunk_impoundment_mortality}{Mortality incurred by Atlantic salmon
smolts during migration through the Mattaceunk (Weldon) Dam impoundment. 
Default usage is to inherit these values from the \code{link{run_dia}} 
function.}

\item{n_dams}{A dataframe with structure identical to the built-in
\code{link{n_dams}} data set.}

\item{indirect_latent_mortality}{Indirect, latent mortality incurred by 
Atlantic salmon smolts at each dam passed. Default usage is to inherit this value from the 
\code{link{run_dia}} function.}

\item{p_female}{Proportion of females in population. Default usage is to 
inherit this value from the \code{link{run_dia}} function.}
}
\value{
A named list with four elements including total number of hatchery and
wild smolts exiting the Penobscot River (elements 1 & 2) as well as the 
proportion of wild and hatchery smolts exiting the river from each natal PU.
}
\description{
Run the downstream passage routine for Atlantic salmon Dam
Impact Analysis (DIA) model v67.
}
\references{
Nieland et al. (2015, 2020).
}
\seealso{
run_dia, n_dams
}
