% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smolt_stocking.R
\name{smolt_stocking}
\alias{smolt_stocking}
\title{Smolt stocking by production unit}
\usage{
smolt_stocking(n_stocked = 545000, stocking_proportions = NULL)
}
\arguments{
\item{n_stocked}{Number of hatchery-reared 2-year smolts to be stocked.}

\item{stocking_proportions}{Proportion of total to be stocked in each PU. The
values must sum to 1.}
}
\value{
A vector with 15 elements within which positions correspond to PUs
as ordered in the \code{link{stocking_default}} dataset. The default (`NULL`)
corresponds to default values used in Nieland et al. (2015, 2020).
}
\description{
Create smolt stocking numbers in each production unit (PU) based
on starting number of smolts and proportion of smolts to be stocked in each
PU.
}
\references{
Nieland et al. (2015, 2020).
}
