% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stillwater_use}
\alias{stillwater_use}
\title{Stillwater Use}
\format{
A dataframe with 5,000 observations of 2 variables
\describe{
  \item{\code{cum_prob}}{Cumulative probability for proportion}
  \item{\code{proportion}}{Proportion of smolts using Stillwater Branch}
}
}
\source{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.
}
\usage{
stillwater_use
}
\description{
Stillwater use table from the Atlantic salmon 
Dam Impact analysis model v67. Cumulative probabilities of proportional
use of Stillwater Branch by outmigrating Atlantic salmon smolts in the 
Penobscot River, ME. This dataset was updated in 2020 following changes
associated with the Penobscot River Restoration Project 
(Nieland et al. 2020).
}
\keyword{datasets}
