% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stillwater_use_old}
\alias{stillwater_use_old}
\title{Stillwater Use (old)}
\format{
A dataframe with 5,000 observations of 2 variables
\describe{
  \code{cum_prob } Cumulative probability for proportion \cr \cr
  \code{proportion } Proportion of smolts using Stillwater Branch \cr \cr
}
}
\source{
Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.
}
\usage{
stillwater_use_old
}
\description{
Stillwater use table from the Atlantic salmon 
Dam Impact analysis model v67. Previous values used for 
Cumulative probabilities of proportional use of Stillwater Branch by 
outmigrating Atlantic salmon smolts in the Penobscot River, ME.
}
\keyword{datasets}
