% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocking_default}
\alias{stocking_default}
\title{Default stocking proportions by production unit}
\format{
A dataframe with 15 observations of 2 variables
\describe{
  \code{PU } Production unit \cr \cr
  \code{proportion_stocked } Proportion of hatchery smolts stocked in PU \cr \cr
}
}
\source{
Nieland et al. (2020)
}
\usage{
stocking_default
}
\description{
Default stocking proportions from the Hatchery worksheet
in the Atlantic salmon Dam Impact Analysis model v67.
}
\keyword{datasets}
