% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straying_locations}
\alias{straying_locations}
\title{Straying locations}
\format{
A dataframe with 15 observations of 16 variables.
\describe{
  \code{Natal_PU }{Natal production unit as smolt} \cr \cr
  \code{PU_01, ..., PU14 } Destination production unit as adult \cr \cr
}
}
\source{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.
}
\usage{
straying_locations
}
\description{
Straying matrix from the Atlantic salmon Dam 
Impact Analysis model v67. Straying locations of adult Atlantic 
salmon based on behavioral patterns.
}
\keyword{datasets}
