% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straying_matrix}
\alias{straying_matrix}
\title{Straying matrix}
\format{
A matrix with 15 observations of 16 variables.
}
\source{
Nieland et al. (2013, 2015, 2020)
}
\usage{
straying_matrix
}
\description{
Numeric columns only from \code{\link{straying_locations}}
}
\keyword{datasets}
