% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream_inefficiency}
\alias{upstream_inefficiency}
\title{Upstream inefficiency of adult migrants}
\format{
A dataframe with 15 observations of 19 variables
\describe{
  \code{dam_failed_to_pass } The dam the adults fail to pass \cr \cr
  \code{p_die } The proportion of failed migrants that die \cr \cr
  \code{p_return_to_sea } The proportion of failed migrants that return to sea \cr \cr
  \code{p_go_elsewhere } The proportion going elsewhere \cr \cr
  \code{PU_01...PU14 } The destination PUs for failed migrants at each dam with proportion going to each \cr \cr
}
}
\source{
Nieland et al. (2013, 2015, 2020)
}
\usage{
upstream_inefficiency
}
\description{
Upstream inefficiency table from the Atlantic salmon 
Dam Impact analysis model v67. When adult Atlantic salmon do not 
pass a dam, a proportion of those fish die, a proportion return 
to sea, and a proportion go elsewhere to attempt to spawn.
}
\keyword{datasets}
