% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downstream_sebec}
\alias{downstream_sebec}
\title{Downstream data for Sebec}
\format{
A dataframe with 745 observations of 4 variables
\describe{
  \item{\code{flow_cfs}}{Value of flow (discharge)}
  \item{\code{cumulative_prob}}{Cumulative flow probability}
  \item{\code{flow_prob}}{Probability of individual flow value}
  \item{\code{smolt_surv_prob}}{Smolt survival probability given flow}
}
}
\source{
Amaral S, Fay C, Hecker G, Perkins N. 2012. Atlantic salmon survival 
estimates at mainstem hydroelectric projects on the Penobscot River. 
Holden (MA): Alden Research Laboratory, Inc. Phase 3 Final Report. 

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.
}
\usage{
downstream_sebec
}
\description{
Alden research estimates of downstream survival from the 
Atlantic salmon Dam Impact analysis model v67 for Sebec Dam.
Annual survival estimates from a variety of acoustic- and 
radio-telemetry studies were correlated with annual, in-river discharge data
from hydropower facilities by Amaral et al. (2012) to result in cumulative 
distribution functions from which smolt survival can be drawn 
probabilistically. This is the approach used in the list of
\code{downstream} passage rates for \code{sebec} in the 
\code{\link{run_dia}} function when \code{sebec = NA}.
}
\keyword{datasets}
