% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_stage_survival}
\alias{life_stage_survival}
\title{Life-stage survival rates}
\format{
A dataframe with 5 observations of 6 variables
\describe{
  \code{rate } Vital rate of interest
  \code{means } Mean of vital rate
  \code{medians } Median of vital rate
  \code{sds } Standard deviation of vital rate
  \code{mins } Minimum value (where applicable)
  \code{maxes } Maximum value (where applicable)
}
}
\source{
Nieland et al. (2013, 2015, 2020)
}
\usage{
life_stage_survival
}
\description{
Life Stage Survival table from the Atlantic salmon 
Dam Impact analysis model v67. Distributional parameters for year-specific
number of eggs produced per female, egg-to-smolt survival rate, 
in-river mortality rate, and marine survival rate.
}
\keyword{datasets}
