% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_stage_survival}
\alias{life_stage_survival}
\title{Life-stage survival rates}
\format{
A dataframe with 5 observations of 6 variables
\describe{
  \code{rate } Vital rate of interest \cr \cr
  \code{means } Mean of vital rate \cr \cr
  \code{medians } Median of vital rate \cr \cr
  \code{sds } Standard deviation of vital rate \cr \cr
  \code{mins } Minimum value (where applicable) \cr \cr
  \code{maxes } Maximum value (where applicable) \cr \cr
}
}
\source{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.
}
\usage{
life_stage_survival
}
\description{
Life Stage Survival table from the Atlantic salmon 
Dam Impact analysis model v67. Distributional parameters for year-specific
number of eggs produced per female, egg-to-smolt survival rate, 
in-river mortality rate, and marine survival rate.
}
\keyword{datasets}
