% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_egg_to_smolt_survival.R
\name{make_egg_to_smolt_survival}
\alias{make_egg_to_smolt_survival}
\title{Simulate egg-to-smolt survival}
\usage{
make_egg_to_smolt_survival(n = 1, a = 0, b = 0.023, mean = NULL, sd = NULL)
}
\arguments{
\item{n}{Number of samples}

\item{a}{Lower limit of truncated normal}

\item{b}{Upper limit of truncated normal}

\item{mean}{Mean of distribution. When `mean = NULL` (default), the mean
value is parameterized from the built-in \code{\link{life_stage_survival}}
dataset. Values are based on available studies cited in Nieland and 
Sheehan (2020).}

\item{sd}{Standard deviation of distribution. When `mean = NULL` (default), 
the mean value is parameterized from the built-in
\code{\link{life_stage_survival}} dataset (Nieland and Sheehan 2020).}
}
\value{
A vector of length `n`
}
\description{
Simulate egg-to-smolt survival from the built-in
\code{\link{life_stage_survival}} dataset using a truncated normal 
distribution (Nieland and Sheehan 2020).
}
\references{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.
}
