% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs_per_female.R
\name{make_eggs_per_female}
\alias{make_eggs_per_female}
\title{Simulate eggs per female}
\usage{
make_eggs_per_female(n = 1, a = 4000, b = 12000, mean = NULL, sd = NULL)
}
\arguments{
\item{n}{Number of samples}

\item{a}{Lower limit of truncated normal}

\item{b}{Upper limit of truncated normal}

\item{mean}{Mean of distribution. When `mean = NULL` (default), the mean
value is parameterized from the built-in \code{\link{life_stage_survival}}
dataset.}

\item{sd}{Standard deviation of distribution. When `mean = NULL` (default), 
the mean value is parameterized from the built-in
 \code{\link{life_stage_survival}} dataset.}
}
\value{
A vector of length `n`
}
\description{
Simulate number of eggs per adult female based on data
collected at the US Fish and Wildlife Service Craig Brook National Fish
Hatchery.
}
\references{
Nieland et al. (2020)
}
