% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_marine_s.R
\name{make_marine_s}
\alias{make_marine_s}
\title{Simulate marine survival}
\usage{
make_marine_s(n = 1, a = 0, b = 0.25, mean = NULL, sd = NULL, hatchery)
}
\arguments{
\item{n}{Number of samples}

\item{a}{Lower limit of truncated normal}

\item{b}{Upper limit of truncated normal}

\item{mean}{Mean of distribution. When `mean = NULL` (default), the mean
value is parameterized from the built-in \code{\link{life_stage_survival}}
dataset.}

\item{sd}{Standard deviation of distribution. When `mean = NULL` (default), 
the mean value is parameterized from the built-in
 \code{\link{life_stage_survival}} dataset.}

\item{hatchery}{A logical indicating whether the estimate is
for hatchery-reared fish (`TRUE`) or wild fish (`FALSE`).}
}
\value{
A vector of length `n`
}
\description{
Simulate marine survival from the built-in
\code{\link{life_stage_survival}} dataset using a truncated normal 
distribution.
}
\references{
Nieland et al. (2020)
}
