% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inf.R
\name{prep_general_inputs}
\alias{prep_general_inputs}
\title{Prepare input data that will be used in all scenarios}
\usage{
prep_general_inputs(
  rd.exist.file,
  pad.exist.file,
  oil.av.file,
  cost.map.file,
  d2cpf = 35000,
  wd.loc,
  path.in,
  proj.info
)
}
\arguments{
\item{rd.exist.file}{Character string specifying the file name for the
shapefile depicting existing roads as lines.}

\item{pad.exist.file}{Character string specifying the file name for the
shapefile depicting existing central processing facility (CPF) and satellite
pads as points.}

\item{oil.av.file}{Character string specifying the file name for the relative
estimated undiscovered oil raster.}

\item{cost.map.file}{Character string specifying the file name for the cost map
raster, which has all water areas coded as 0 and all land as 1.}

\item{d2cpf}{Numeric value indicating the minimum allowed distance (m) between
two central processing facilities (CPFs). Defaults to 35,000 m.}

\item{wd.loc}{Character string identifying the base directory containing both
input and output data folders.}

\item{path.in}{Character string identifying the relative path from the base
directory (\code{wd.loc}) to where input data are stored. Defaults to an
Input_Data folder within the base folder.}

\item{proj.info}{Desired projection string in EPSG code format (\code{"EPSG:XXXX"}),
common to all spatial objects in the analysis.}
}
\value{
List containing input data used for infrastructure simulation across
all scenarios. This includes the land/water \code{SpatRaster} (\code{cost_map}),
road resistance \code{SpatRaster} that makes development across water possible
(\code{cost_water}), \code{SpatRaster} indicating relative estimated undiscovered
oil (\code{oil_av}), \code{SpatRaster} depicting the infrastructure exclusion zone
around the Willow CPF (\code{willow_cpf_ras}), \code{sf} \code{POINT} object depicting
existing CPF and satellite gravel pads (\code{pad_exist}), \code{sf} \code{LINESTRING}
object depicting existing roads (\code{rd_exist}), data.frame with xy
coordinates of existing roads (\code{exist_coords}), and \code{sf} \code{POINT} object
depicting the Willow CPF location (\code{willow_spdf}). Many of these feed
into \code{\link{prep_scenario_inputs}}.
}
\description{
\code{prep_general_inputs} loads and prepares input data that is common to all
scenarios to be evaluated, like existing infrastructure layers, relative
estimated undiscovered oil, and the land/water cost map.
}
