% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dia.R
\name{run_dia}
\alias{run_dia}
\title{Run the Dam Impact Analysis}
\usage{
run_dia(
  n_wild = 31,
  n_hatchery = 306,
  stocking = 1,
  n_stocked = 545000,
  upstream = list(medway = 1, mattaceunk = 1, west_enfield = 1, upper_dover = 1,
    browns_mills = 1, sebec = 1, milo = 1, howland = 1, lowel = 1, stillwater = 1,
    milford = 1, great_works = 1, orono = 1, veazie = 1, frankfort = 1),
  downstream = list(medway = 0, mattaceunk = 1, west_enfield = 0.96, upper_dover =
    0.9215, browns_mills = NA, sebec = NA, milo = NA, howland = 1, lowell = NA,
    stillwater = 0.96, milford = 0.96, great_works = 1, orono = 0.96, veazie = 1,
    frankfort = NA),
  mattaceunk_impoundment_mortality = 0.072,
  p_stillwater = NULL,
  indirect_latent_mortality = 0.06,
  p_female = 0.6,
  new_or_old = "new",
  marine_s_hatchery = NULL,
  marine_s_wild = NULL,
  straying_matrix = NULL,
  p_mainstem_up = 1,
  n_broodstock = 150
)
}
\arguments{
\item{n_wild}{Number of starting wild adult salmon.}

\item{n_hatchery}{Number of starting hatchery adult salmon.}

\item{stocking}{Number indicating whether hatchery stocking is on (`1`) or
off (`0`). Will be a vector corresponding to `n_generations` at some point.}

\item{n_stocked}{Number of fish to be stocked. Will also be a vector 
corresponding to `n_generations` in length at some point.}

\item{upstream}{Upstream passage efficiency of adult Atlantic salmon through
dams.}

\item{downstream}{Downstream survival of Atlantic salmon smolts through dams.
The default values are based on the most recent values used by Nieland et al. 
(2020) based on standards established in the the NOAA Species Protection Plan
following implementation of the Penobscot River Restoration Project.
The default `NULL` randomly samples correlated survival rates of smolts for 
each dam based on cumulative flow probabilities and associated empirical 
survival rates (Nieland et al. 2013, 2020).}

\item{mattaceunk_impoundment_mortality}{Mortality incurred by Atlantic salmon
smolts during migration through the Mattaceunk (Weldon) Dam impoundment. The
default value is based on Nieland et al. (2020). Based on results of 
Holbrook et al. (2011) and Stich et al. (2015a).}

\item{p_stillwater}{Probability that fish use the Stillwater Branch for 
downstream migration. The default (`NULL`) draws flow-correlated values from
a cumulative distribution of flows with paired estimates of p_stillwater
used by Nieland et al. (2020), based on empirical results in Holbrook et al.
(2006), Stich et al. (2014), and Stich et al. (2015a).}

\item{indirect_latent_mortality}{Indirect, latent mortality incurred by 
Atlantic salmon smolts at each dam passed. The default value is what was used
in Nieland et al. (2020), derived from estimates in Stich et al. (2015b).}

\item{p_female}{Proportion of females in population.}

\item{new_or_old}{A character string indicating whether to use `"new"` 
(Nieland et al. 2013, 2015) or `"old"` (Nieland et al. 2020) flow-correlated
probabilities of p_stillwater as well as flow-correlated survival at 
`milford`, `orono`, and `stillwater` dams.}

\item{marine_s_hatchery}{Numeric indicating marine survival rate for 
post-smolt to adult survival of hatchery outmigrants. The default (`NULL`)
simulates values from a truncated normal distribution using hatchery smolt
survival estimates from the Penobscot River, ME, USA.}

\item{marine_s_wild}{Numeric indicating marine survival rate for 
post-smolt to adult survival of wild outmigrants. The default (`NULL`)
simulates values from a truncated normal distribution using wild smolt
survival estimates from the Narraguagus River, ME, USA.}

\item{straying_matrix}{A dataframe identical in structure to the built-in
\code{\link{straying_locations}} dataset.}

\item{p_mainstem_up}{Probability that fish use the mainstem Penobscot River
(and not Stillwater Branch) for upstream migration around Marsh Island.}

\item{n_broodstock}{Target number of adult returns collected at Milford Dam 
for spawning at US Fish and Wildlife Service Craig Brook National Fish 
Hatchery each year. Broodstock are collected upstream of Milford Dam
in \code{\link{run_upstream_passage}}.}

\item{n_generations}{Number of generations to simulate for simulation. One 
complete generation from egg to returning adult is six years.}
}
\value{
A dataframe containing inputs and outputs. Will add more
of an explanation here once we know what they all are.
}
\description{
Use functions from \code{\link{dia}} to simulate one
generation of 2 sea-winter female Atlantic salmon through time using data
and inputs from Nieland et al. (2013, 2015, 2020) as implemented in `@Risk`.
}
\references{
Holbrook et al. (2006); Nieland et al. (2013, 2015, 2020); Stich 
et al. (2014, 2015a, 2015b).
}
