% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inefficiency.R
\name{run_inefficiency}
\alias{run_inefficiency}
\title{Run adult passage inefficiency}
\usage{
run_inefficiency(n, inefficiency_matrix)
}
\arguments{
\item{n}{A vector of length 15 containing total number of fish that failed to
pass dams and did not die or return to sea.}

\item{inefficiency_matrix}{A matrix identical in structure to the built-in
\code{\link{inefficiency_matrix}} dataset (NMFS 2012; Neiland et al. 2013).}
}
\description{
Adult inefficiency (homing) routine for Atlantic salmon that 
fail to pass a dam but that also do not die or return to sea.
}
\examples{
# Simulate 10,000 starting fish at each dam and
# randomly sample number of failures for each PU (15 of these)

failures = sample(1e4, 15, replace = TRUE)

# Use run_inefficiency to simulate where these 
# failed migrants end up in the watershed based on the 
# default inefficiency_matrix in dia.

run_inefficiency(failures, dia::inefficiency_matrix)

}
\references{
Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.

NMFS (National Marine Fisheries Service). 2012. Biological Opinion regarding 
effects to the endangered GOM DPS for Atlantic salmon (Salmo salar), endangered 
shortnose sturgeon (Acipenser brevirostrum), and the threatened and endangered 
DPSs of Atlantic sturgeon (Acipenser oxyrhynchus oxyrhynchus) from proposed 
FERC license amendments for Black Bear Hydro Partners’ Hydroelectric Projects 
on the Penobscot River in Maine. NMFS Greater Atlantic Regional Office, 
Gloucester, MA.
}
\seealso{
inefficiency_matrix, upstream_inefficiency
}
