% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inefficiency.R
\name{run_inefficiency}
\alias{run_inefficiency}
\title{Run adult passage inefficiency}
\usage{
run_inefficiency(n, inefficiency_matrix)
}
\arguments{
\item{n}{A vector of length 15 containing total number of fish that failed to
pass dams and did not die or return to sea.}

\item{inefficiency_matrix}{A matrix identical in structure to the built-in
\code{\link{inefficiency_matrix}} dataset.}
}
\description{
Adult inefficiency (homing) routine for Atlantic salmon that 
fail to pass a dam but that also do not die or return to sea.
}
\references{
Nieland et al. (2015, 2020).
}
\seealso{
inefficiency_matrix, upstream_inefficiency
}
