% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_straying.R
\name{run_straying}
\alias{run_straying}
\title{Run adult straying routine}
\usage{
run_straying(returns, straying_matrix)
}
\arguments{
\item{returns}{Total number of fish returning to river.}

\item{straying_matrix}{A matrix identical in structure to the built-in
\code{\link{straying_matrix}} dataset.}
}
\description{
Adult straying (homing) routine for Atlantic salmon in
the Dam Impact Analysis (DIA) model v67 during year 6.
}
\references{
Nieland et al. (2015, 2020).
}
\seealso{
straying_matrix, straying_locations
}
