% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_upstream_passage.R
\name{run_upstream_passage}
\alias{run_upstream_passage}
\title{Run upstream migration routine}
\usage{
run_upstream_passage(
  hatchery_returns,
  hatchery_proportions,
  wild_returns,
  wild_proportions,
  upstream_passage,
  straying_matrix,
  p_mainstem_up,
  stocking,
  n_broodstock,
  inefficiency_matrix
)
}
\arguments{
\item{hatchery_returns}{Total number of hatchery smolts exiting river in 
year 4.}

\item{hatchery_proportions}{Proportion of individuals attempting to return
to each production unit (PU in the river). Within the `run_dia()` function,
these values are passed as output from `run_downstream_migration()`.}

\item{wild_returns}{Number of wild-reared smolts exiting river in 
year 4.}

\item{wild_proportions}{Proportion of individuals attempting to return
to each production unit (PU in the river). Within the `run_dia()` function,
these values are passed as output from `run_downstream_migration()`.}

\item{upstream_passage}{Numeric vector with length = 15 corresponding to
dam passage efficiencies at each of 15 dams in the watershed that are used 
to delineate PUs. Default usage is to inherit these values from the 
\code{link{run_dia}} function's `upstream_passage` argument.}

\item{straying_matrix}{A dataframe identical in structure to the built-in
\code{\link{straying_locations}} dataset.}

\item{p_mainstem_up}{Probability that fish use the mainstem Penobscot River
(and not Stillwater Branch) for upstream migration around Marsh Island.}

\item{stocking}{Number indicating whether hatchery stocking is on (`1`) or
off (`0`). Will be a vector corresponding to `n_generations` at some point.}

\item{n_broodstock}{Target number of adult returns collected at Milford Dam 
for spawning at US Fish and Wildlife Service Craig Brook National Fish 
Hatchery each year. Broodstock are collected upstream of Milford Dam
in \code{\link{run_upstream_passage}}.}

\item{inefficiency_matrix}{A matrix identical in structure to the built-in
\code{\link{inefficiency_matrix}} dataset.}
}
\description{
Run the upstream migration routine for adult Atlantic salmon in
the Dam Impact Analysis (DIA) model v67 during year 6.
}
\references{
Nieland et al. (2015, 2020).
}
\seealso{
run_dia, run_downstream_passage
}
