% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stillwater_use}
\alias{stillwater_use}
\title{Stillwater Use}
\format{
A dataframe with 5,000 observations of 2 variables
\describe{
  \code{cum_prob } Cumulative probability for proportion \cr \cr
  \code{proportion } Proportion of smolts using Stillwater Branch \cr \cr
}
}
\source{
Nieland et al. (2020)
}
\usage{
stillwater_use
}
\description{
Stillwater use table from the Atlantic salmon 
Dam Impact analysis model v67. Cumulative probabilities of proportional
use of Stillwater Branch by outmigrating Atlantic salmon smolts in the 
Penobscot River, ME. This dataset was updated in 2020 following changes
associated with the Penobscot River restoration project.
}
\keyword{datasets}
