% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stillwater_use_old}
\alias{stillwater_use_old}
\title{Stillwater Use (old)}
\format{
A dataframe with 5,000 observations of 2 variables
\describe{
  \code{cum_prob } Cumulative probability for proportion
  \code{proportion } Proportion of smolts using Stillwater Branch
}
}
\source{
Nieland et al. (2013, 2015)
}
\usage{
stillwater_use_old
}
\description{
Stillwater use table from the Atlantic salmon 
Dam Impact analysis model v67. Previous values used for 
Cumulative probabilities of proportional use of Stillwater Branch by 
outmigrating Atlantic salmon smolts in the Penobscot River, ME.
}
\keyword{datasets}
