% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straying_locations}
\alias{straying_locations}
\title{Straying locations}
\format{
A dataframe with 15 observations of 16 variables.
\describe{
  \code{Natal_PU }{Natal production unit as smolt}
  \code{PU_01, ..., PU14 } Destination production unit as adult
}
}
\source{
Nieland et al. (2013, 2015, 2020)
}
\usage{
straying_locations
}
\description{
Straying matrix from the Atlantic salmon Dam 
Impact Analysis model v67.Straying locations of adult Atlantic 
salmon based on behavioral patterns.
}
\keyword{datasets}
