% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream_inefficiency}
\alias{upstream_inefficiency}
\title{Upstream inefficiency of adult migrants}
\format{
A dataframe with 15 observations of 15 variables
\describe{
  \item{\code{dam_failed_to_pass}}{The dam the adults fail to pass}
  \item{\code{p_die}}{The proportion of failed migrants that die}
  \item{\code{p_return_to_sea}}{The proportion of failed migrants that return to sea}
  \item{\code{p_go_elsewhere}}{The proportion of failed migrants going elsewhere}
  \item{\code{PU_01...PU14}}{The destination PUs for failed migrants at each dam with proportion of failed migrants going to each}
}
}
\source{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.
}
\usage{
upstream_inefficiency
}
\description{
Upstream inefficiency table from the Atlantic salmon 
Dam Impact Analysis model v67. When adult Atlantic salmon do not 
pass a dam, a proportion of those fish die, a proportion return 
to sea, and a proportion go elsewhere to attempt to spawn. These values are
used in combination with the \code{\link{straying_matrix}} to spatially
distribute returning adult spawners in each generation of simulations.
}
\keyword{datasets}
